Follow review comments
